ankitools_project/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ankitools"
version = "0.1.0"
description = "A library and CLI for automating Anki tasks and enhancing note creation with AI."
readme = "README.md"
requires-python = ">=3.9" # Specify a reasonable minimum Python version
license = {text = "MIT"}
authors = [
  {name = "Cline (AI Software Engineer)", email = "cline@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
    "Topic :: Education",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.20.0",
    "google-generativeai>=0.5.0", # Check for latest appropriate version
    "python-dotenv>=0.20.0",
    "pyphen>=0.14.0",
    "typer[all]>=0.9.0", # [all] includes rich for nice CLI output
]

[project.urls]
Homepage = "https://github.com/example/ankitools" # Placeholder
Repository = "https://github.com/example/ankitools" # Placeholder

[project.scripts]
ankitools = "ankitools_lib.cli.main:app" # Assuming Typer app is in cli.main

[tool.setuptools.packages.find]
where = ["."] # find packages in the current directory
include = ["ankitools_lib*"]  # include the ankitools_lib package and its submodules
exclude = ["tests*"] # exclude tests package
